[trading-composition][Kotlin] Move contract tests to the same package as the contract
