+++ kotlin/trading-composition/adapters/exposed/build.gradle.kts
@@ -7,4 +7,5 @@ dependencies {
     implementation(libs.bundles.exposed)
     implementation(libs.postgresql)
     testImplementation(libs.bundles.testcontainers)
+    testImplementation(testFixtures(project(":trading-composition:domain")))
 }
