+++ kotlin/trading-composition/domain/build.gradle.kts
@@ -1,3 +1,9 @@
 plugins {
     id("kotlin-common-conventions")
+    id("java-test-fixtures")
+}
+
+dependencies {
+    testFixturesImplementation(kotlin("test"))
+    testFixturesImplementation(kotlin("test-junit5"))
 }
